students/04/mpuizina/requirements.txt
marimo
requests
python-dotenv
google.generativeai
beautifulsoup4